refactor(shortcut-manager): move shortcut hook to dedicated directory
