+++ src/features/ui-manager/components/tool-bar/UiControlsButton/UiControlsButton.tsx
@@ -1,6 +1,6 @@
 import ToggleToolBarButton from "@/components/tool-bar/ToolBarToggleButton";
 import { useUiControlsPanelStore } from "@/features/ui-manager/stores/uiControlsPanelStore";
-import { useShortcut } from "@/hooks/useShortcut";
+import { useShortcut } from "@/hooks/shortcut-manager/useShortcut";
 
 export default function UiControlsButton() {
   const isVisible = useUiControlsPanelStore((state) => state.isVisible);
