+++ src/hooks/shortcut-manager/useShortcut.ts
